Add files via upload
Here is the complete code. The bounding box code is device specific as of now. i.e. (393x393 pixels) wrto my device. The bounding box coordinates are getting sent and displaying correctly on the density map...for the images taken in Potrait mode in camera...also gallery images displays correctly....